refactor(gateway): move ssh server impl into ssh/server.rs (<1000 lines)
ssh.rs was 1021 lines. Convert it to an ssh/ directory module and lift the
inline `mod server { ... }` block (the russh-backed acceptor, auth handler,
and SshTransport) into ssh/server.rs. ssh/mod.rs (453 lines) keeps the wire
codec, SshConfig, authorized-clients handling, and the stdio subsystem
transport. Re-export and behavior unchanged (pub use server::*).

Every file in the gateway crate is now under 1000 lines.
